backend/requirements.txt
flask
flask-cors
mysql-connector-python
gunicorn==21.2.0
